Add component for TransitionInfoTab
